+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - `richmenu unlink`
 ### Changed
  - Re-arrange columns in `richmenu list` and add `name` column
+ - Update dependencies
 
 ## [0.5.0]
 ### Added